Dynamic UIPicker data source
Updated the view controller to build dynamically off of data and not have 2 rows with pickers each time. Now it’s dependent on the number of items the NSDataDetector detects. Also, removed the header and footer cells to the table view making it much more appealing. The constants were updated so there’s now only the starting row and not two constants which store where the pickers are.